requirements.txt
flask
google-generativeai
PyPDF2
python-dotenv